requirements.txt
pymongo==4.3.3
flask==1.1.2
Flask-SocketIO==5.3.2
pandas==1.5.3
validator==0.7.1